--- utilities/data/uno.json
@@ -1,102 +1,0 @@
-{
-  "Uno_Backside": {
-    "Location": ""
-  },
-  "Uno_Wildcards": [
-    "Wild",
-    "Block Wild",
-    "Reverse Wild",
-    "Replay Wild",
-    "# Wild",
-    "Recycle",
-    "Eye",
-    "Shield",
-    "Joker",
-    "+2 Wild",
-    "+4 Wild",
-    "+6 Wild",
-    "+8 Wild",
-    "+10 Wild",
-    "-1 Wild",
-    "-2 Wild",
-    "2x Wild",
-    "Divide 2 Wild",
-    "+4T Wild",
-    "++4 Wild",
-    "D4",
-    "D6",
-    "D8",
-    "D12",
-    "D20",
-    "Trade",
-    "Steel",
-    "Gift",
-    "777",
-    "NO U"
-  ],
-  "4-Color-Uno": {
-    "Wild_Override": {
-      "Wild": 4,
-      "Eye": 2,
-      "Shield": 4,
-      "+2 Wild": 2,
-      "+4 Wild": 2,
-      "Trade": 2
-    },
-    "Uno-Configuration": [
-      {
-        "Name": "0",
-        "Amount": 1
-      },
-      {
-        "Name": "1-9",
-        "Amount": 3,
-        "Speed": false
-      },
-      {
-        "Name": "1-9",
-        "Amount": 1,
-        "Speed": true
-      },
-      {
-        "Name": "Block",
-        "Amount": 4
-      },
-      {
-        "Name": "Reverse",
-        "Amount": 4
-      },
-      {
-        "Name": "Replay",
-        "Amount": 4
-      },
-      {
-        "Name": "#",
-        "Amount": 1
-      },
-      {
-        "Name": "+1",
-        "Amount": 4
-      },
-      {
-        "Name": "+2",
-        "Amount": 3
-      },
-      {
-        "Name": "+4",
-        "Amount": 1
-      },
-      {
-        "Name": "-1",
-        "Amount": 1
-      },
-      {
-        "Name": "Colorless",
-        "Amount": 1
-      },
-      {
-        "Name": "Wild"
-      }
-    ]
-  }
-}
